pyproject.toml
[tool.poetry]
name = "prometeeos-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "apollo", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
langchain = "^0.1.16"
langchain-community = "^0.0.34"
langchainhub = "^0.1.15"
langchain-openai = "^0.1.3"
langchain-pinecone = "^0.1.0"
python-dotenv = "^1.0.1"
pypdf = "^4.2.0"
azure-search-documents = "^11.4.0"
azure-identity = "^1.16.0"
pandas = "^2.2.2"
streamlit = "^1.33.0"
fpdf = "^1.7.2"
pymupdf = "^1.24.2"
langchain-core = "^0.1.52"
requests = "^2.32.2"
beautifulsoup4 = "^4.12.3"
ipykernel = "^6.29.4"
pdfkit = "^1.0.0"
wkhtmltopdf = "^0.2"
fpdf2 = "^2.7.9"
python-docx = "^1.1.2"


[tool.poetry.group.dev.dependencies]
alembic = "^1.13.1"
ruff = "^0.1.7"
pytest = "^7.4.3"
pre-commit = "^3.6.0"
commitizen = "^3.13.0"
doit = "^0.36.0"
types-passlib = "^1.7.7.20240106"


[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "poetry"
update_changelog_on_bump = true
major_version_zero = true
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4
# Assume Python 3.11
target-version = "py311"
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = ["I"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
# Like Black, indent with spaces, rather than tabs.
indent-style = "space"
# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false
# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
disallow_incomplete_defs = true
disallow_untyped_defs = true

[tool.dagster]
module_name = "dags"
